refactor(tx-flow): move TxChecks to tx-flow/features folder
