+++ apps/web/src/components/tx-flow/features/TxChecks/index.tsx
@@ -2,16 +2,8 @@ import { useContext } from 'react'
 import { SafeTxContext } from '@/components/tx-flow/SafeTxProvider'
 import { TxFlowContext } from '@/components/tx-flow/TxFlowProvider'
 import useIsCounterfactualSafe from '@/features/counterfactual/hooks/useIsCounterfactualSafe'
-import TxChecksComponent from '@/components/tx/SignOrExecuteForm/TxChecks'
-import { SlotName, withSlot } from '../slots'
-
-const TxChecks = () => {
-  const { safeTx } = useContext(SafeTxContext)
-
-  if (!safeTx) return false
-
-  return <TxChecksComponent transaction={safeTx} />
-}
+import TxChecks from './TxChecks'
+import { SlotName, withSlot } from '../../slots'
 
 const useShouldRegisterSlot = () => {
   const { isRejection } = useContext(TxFlowContext)
@@ -22,7 +14,13 @@ const useShouldRegisterSlot = () => {
 }
 
 const TxChecksSlot = withSlot({
-  Component: TxChecks,
+  Component: () => {
+    const { safeTx } = useContext(SafeTxContext)
+
+    if (!safeTx) return false
+
+    return <TxChecks transaction={safeTx} />
+  },
   slotName: SlotName.Feature,
   id: 'txChecks',
   useSlotCondition: useShouldRegisterSlot,
