+++ apps/web/src/components/tx/ReviewTransaction/ReviewTransactionContent.tsx
@@ -29,7 +29,7 @@ import ConfirmationView from '../confirmation-views'
 import { TxNoteForm, encodeTxNote, trackAddNote } from '@/features/tx-notes'
 import { SignerForm } from '../SignOrExecuteForm/SignerForm'
 import UnknownContractError from '../SignOrExecuteForm/UnknownContractError'
-import TxChecks from '../SignOrExecuteForm/TxChecks'
+import TxChecks from '@/components/tx-flow/features/TxChecks/TxChecks'
 import { Button, CardActions, CircularProgress, Stack } from '@mui/material'
 import BatchButton from '@/components/tx-flow/actions/Batching/BatchButton'
 import { TxModalContext } from '@/components/tx-flow'
