+++ apps/web/src/components/tx/SignOrExecuteForm/styles.module.css
@@ -38,15 +38,3 @@
 [data-theme='dark'] .execute svg {
   color: var(--color-success-dark);
 }
-
-
-.mobileTxCheckMessages,
-.mobileTxCheckMessages:empty {
-  display: none;
-}
-
-@media (max-width: 899.95px) {
-  .mobileTxCheckMessages {
-    display: block;
-  }
-}
